pyproject.toml
[tool.poetry]
name = "yeastar-sms-bridge"
version = "0.1.0"
description = "A django app that provides additional reliability and admin features for SMS on the Yeastar TG series."
authors = ["Adedayo Ayeni <dedayoa@gmail.com>"]
license = "UNLICENSED"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.3"
django-timezone-field = "^4.1.2"
psycopg2 = "^2.8.6"
django-environ = "^0.4.5"
django-phonenumber-field = "^5.1.0"
sms-counter = {git = "https://github.com/dedayoa/sms-counter-python.git"}
djangorestframework = "^3.12.4"
phonenumbers = "^8.12.23"
requests = "^2.25.1"
gunicorn = "^20.1.0"
whitenoise = "^5.2.0"

[tool.poetry.dev-dependencies]
pytest-django = "^4.3.0"
pytest-mock = "^3.6.1"
pytest-factoryboy = "^2.1.0"
django-debug-toolbar = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
